Date range in mixin
