+++ src/components/finder/filters/TimeFilter.vue
@@ -27,7 +27,7 @@
           transition="scale-transition"
           offset-y
           max-width="290px"
-          min-width="290px"
+          min-width="280px"
           >
 
             <template v-slot:activator="{ on, attrs }">
